+++ domain/src/main/java/com/owusu/cryptosignalalert/domain/models/PriceTargetDomain.kt
@@ -11,7 +11,7 @@ data class PriceTargetDomain(
     val atlDate: String? = null,
     val circulatingSupply: Double? = null,
     val currentPrice: Double? = null,
-    val fullyDilutedValuation: Long? = null,
+    val fullyDilutedValuation: Double? = null,
     val high24h: Double? = null,
     val image: String? = null,
     val lastUpdated: String? = null,
